+++ examples/example-esp32s3/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `UsbMidiEventPacket::as_message_bytes` function.
 - `UsbMidiEventPacket::as_raw_bytes` function.
 - `UsbMidiEventPacket::to_raw_bytes` function.
+- `UsbMidiEventPacket::from_message_bytes` function.
 
 ### Changed